requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
idna==3.4
psycopg2-binary==2.9.6
requests==2.31.0
urllib3==1.26.16
